GUI/requirements.txt
PySDL2==0.9.16
pysdl2-dll==2.28.0
